part 14: final render

+++ RTinOneWeekend/src/Utils.h
@@ -1,23 +1,30 @@
 #pragma once
 
-#include <cmath>
+//#include <cmath>
 #include <iostream>
 #include <limits>
 #include <memory>
 #include <vector>
 #include <glm/glm.hpp>
 #include <cstdlib>
-
-#include "Interval.h"
+#include <Walnut/Random.h>
 
 // can replace with FLT_MAX
 const float infinity = std::numeric_limits<float>::infinity();
-const float pi = (float)3.1415926535897932385;
+const float pi = 3.1415926535897932385f;
+
 
 namespace Utils {
+	// can replace with glm::radians(float degrees)
 	inline float DegreesToRadians(float degrees) {
 		return degrees * pi / 180.0f;
 	}
+
+	//bool NearZero(glm::vec3& v) {
+	//	float s = 1e-8f;
+	//	return (std::fabs(v[0]) < s) && (std::fabs(v[1]) < s) && (std::fabs(v[2]) < s);
+	//}
+
 	/*
 	inline float RandomFloat() { return rand() / (RAND_MAX + 1.0); }
 	inline float RandomFloat(float min, float max) { return min + (max - min) * RandomFloat(); }
@@ -45,5 +52,4 @@ namespace Utils {
 		}
 	}
 	*/
-
 }
